requirements.txt
chainlit
google-generativeai
python-dotenv
-e .
